Added custom exceptions

+++ src/main/java/com/ruvalcaba/quizapplication/service/QuestionService.java
@@ -1,14 +1,15 @@
 package com.ruvalcaba.quizapplication.service;
 
 import com.ruvalcaba.quizapplication.DAO.QuestionDAO;
+import com.ruvalcaba.quizapplication.exception.QuestionInvalidFormatException;
+import com.ruvalcaba.quizapplication.exception.QuestionNotFoundException;
 import com.ruvalcaba.quizapplication.model.QuestionModel;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 
-import java.util.ArrayList;
 import java.util.List;
+import java.util.Objects;
+import java.util.stream.Stream;
 
 @Service // or @Component
 public class QuestionService {
@@ -20,38 +21,31 @@ public class QuestionService {
         return questionDAO.findAll();
     }
 
-    public ResponseEntity<List<QuestionModel>> getQuestionsByCategory(String category) {
-        try{
-            return new ResponseEntity<>(questionDAO.findByCategory(category), HttpStatus.OK);
-        } catch(Exception e){
-            e.printStackTrace();
-        }
-        return new ResponseEntity<>(new ArrayList<>(), HttpStatus.BAD_REQUEST);
+    public List<QuestionModel> getQuestionsByCategory(String category) {
+        if(questionDAO.findByCategory(category).isEmpty())
+            throw new QuestionNotFoundException("No questions match the specified category");
+        return questionDAO.findByCategory(category);
     }
 
     public List<QuestionModel> getQuestionsByDifficulty(String difficulty) {
+        if(questionDAO.findByDifficultyLevel(difficulty).isEmpty())
+            throw new QuestionNotFoundException("No questions match the specified difficulty");
         return questionDAO.findByDifficultyLevel(difficulty);
     }
 
-    public ResponseEntity<String> addQuestion(QuestionModel question) {
-        try{
-            questionDAO.save(question);
-            return new ResponseEntity<>("Success", HttpStatus.CREATED);
-        } catch(Exception e){
-            e.printStackTrace();
-        }
-        return new ResponseEntity<>("Error", HttpStatus.BAD_REQUEST);
+    public String addQuestion(QuestionModel q) {
+        if(q.getCategory()==null || q.getQuestionTitle()==null || q.getCorrectAnswer()==null ||
+                q.getDifficultyLevel()==null || q.getOption1()==null || q.getOption2()==null ||
+                q.getOption3()==null || q.getOption4()==null)
+            throw new QuestionInvalidFormatException("Invalid format, some fields are null");
 
+        return "Question added successfully";
     }
 
-    public ResponseEntity<String> deleteQuestion(Integer id) {
-        try{
-            questionDAO.deleteById(id);
-            return new ResponseEntity<>("Question" + id + "deleted succesfully", HttpStatus.OK);
-        }catch (Exception e){
-            e.printStackTrace();
-        }
-        return new ResponseEntity<>("Error", HttpStatus.BAD_REQUEST);
-
+    public String deleteQuestion(Integer id) {
+        if(questionDAO.findById(id).isEmpty())
+            throw new QuestionNotFoundException("The question with specified ID doesn't exists");
+        questionDAO.deleteById(id);
+        return "Question with id=" + id + " deleted successfully";
     }
 }
